Commented and simplified encoding
