+++ tnreason/encoding/__init__.py
@@ -1,10 +1,14 @@
-from tnreason.encoding.formulas import create_formulas_cores, create_raw_formula_cores, get_formula_color, \
-    create_headCore, create_evidence_cores, is_subexpression
-from tnreason.encoding.constraints import create_constraints
-from tnreason.encoding.auxiliary import create_emptyCoresDict, get_atoms, get_all_atoms
-from tnreason.encoding.storage import save_as_yaml, load_from_yaml
-from tnreason.encoding.neurons import create_neuron, create_architecture, find_atoms, find_selection_dimDict, \
+from tnreason.encoding.formulas_to_cores import create_formulas_cores, create_raw_formula_cores, get_formula_color, \
+    create_head_core, create_evidence_cores, get_atoms, get_all_atoms
+from tnreason.encoding.categoricals_to_cores import create_categorical_cores
+from tnreason.encoding.auxiliary_cores import create_trivial_cores
+from tnreason.encoding.neurons_to_cores import create_neuron, create_architecture, find_atoms, find_selection_dimDict, \
     create_solution_expression
-from tnreason.encoding.data import create_data_cores
+from tnreason.encoding.data_to_cores import create_data_cores
 
-from tnreason.encoding.formulas import headCoreSuffix
+## Core Suffix Nomenclature used in other subpackages
+from tnreason.encoding.formulas_to_cores import headCoreSuffix
+
+from tnreason.encoding.storage import save_as_yaml, load_from_yaml
+
+
