+++ tnreason/encoding/truth_tables.py
