+++ tnreason/knowledge/deductive.py
@@ -45,8 +45,8 @@ class InferenceProvider:
     def query(self, variableList, evidenceDict={}):
         return engine.contract(method=self.contractionMethod, coreDict={
             **self.distribution.create_cores(),
-            **encoding.create_emptyCoresDict([variable for variable in variableList if
-                                              variable not in self.distribution.atoms and variable not in evidenceDict]),
+            **encoding.create_trivial_cores([variable for variable in variableList if
+                                             variable not in self.distribution.atoms and variable not in evidenceDict]),
             **encoding.create_evidence_cores(evidenceDict),
         }, openColors=variableList).normalize()
